crud purchase order finalized

+++ src/main/java/com/rigo/local/storage/localStorageProyect/domain/repositories/SupplierRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 public interface SupplierRepository extends JpaRepository<SupplierEntity, Long> {
 
     SupplierEntity findByName(String name);
-    SupplierEntity findByPhone(int phone);
+    SupplierEntity findByPhone(Long phone);
     SupplierEntity findByAddress(String address);
     SupplierEntity findByEmail(String email);
 }
